pyproject.toml
[project]
name = "jonja"
version = "1.0.0"
description = "Simple jinja-based texts and objects render"
authors = [
    {name = "lightmanLP", email = "liteman1000@gmail.com"}
]
license = "MIT"
readme = "README.md"
keywords = ["jinja2", "template", "text"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed"
]
requires-python = ">=3.10,<4"
dependencies = [
    "jinja2>=3.1.6",
]

[project.urls]
repository = "https://github.com/UT1C/jonja"

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "ruamel-yaml>=0.19.1",
    "typing-extensions>=4.15.0",
]
